fix bazel build, separate graph functionality from common
